Fix build for homebrew GCC
Recent versions of XCode Command Line Tools result in failures with
homebrew GCC@9 due to conflicting definitions. This is a result of the
use of Clang specific `__has_builtin` in the XCode CLT SDK
`stdlib.h`[1].

`stdio.h` is included before `stdlib.h` to provide a definition for
`__has_builtin`.

[1]: https://github.com/Homebrew/homebrew-core/issues/40676