Adapt test streams from Beast for test socket
Beast is a library that builds on top of asio and provides test
stream[1] for use in unit tests. However, Beast requires boost and does
not work with asio standalone[2]. As the only functionality we desire
are the test streams, this commit adapts the test stream for use in
testing Connection, and only retains the methods to satisfy the
AsyncReadStream concept.

The additional constraints due to AsyncSocketStream are also implemented
on the test socket class.

[1]: https://github.com/boostorg/beast/blob/boost-1.73.0/include/boost/beast/_experimental/test/stream.hpp
[2]: https://github.com/boostorg/beast/issues/799