Add unit tests for Connection
Test output of `Connection` for given input bytes. In order to allow
testing, `Connection` now takes the `out` and `err` ostreams as
parameters. As a result, the lifetime of `out` and `err` must exceed
that of a `Connection`.

Note that this is the case for `std::cout` and `std::cerr`, and that
`Connection`s are only created within a private member function of
`Server`.